appcode/backend/requirements.txt
azure-functions
fastapi
uvicorn
azure-cosmos
azure-identity
pydantic